Srm tender

+++ app/Http/Controllers/API/TenderProcurementCategoryController.php
@@ -9,6 +9,7 @@ use App\Models\TenderProcurementCategory;
 use Illuminate\Http\Request;
 use App\Repositories\ProcurementCategoryRepository;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Log;
 use Prettus\Validator\Exceptions\ValidatorException;
 
 class TenderProcurementCategoryController extends AppBaseController
@@ -62,18 +63,31 @@ class TenderProcurementCategoryController extends AppBaseController
             return $this->sendError($validator->messages(), 422);
         }
 
-        $procurementCatCodeExist = TenderProcurementCategory::select('id')
+        $procurementCatCodeExist = TenderProcurementCategory::withTrashed()
+            ->select('id', 'deleted_at')
             ->where('code', '=', $input['code'])
             ->where('level', '=', $level)->first();
+
         if (!empty($procurementCatCodeExist)) {
-            return $this->sendError('Procurement code ' . $input['code'] . ' already exists');
+            if(is_null($procurementCatCodeExist['deleted_at'])){
+                return $this->sendError('Procurement code ' . $input['code'] . ' already exists');
+            } else {
+                $errorMessage = 'Procurement code '. $input['code'] . '  already exist, Do you really want to restore this?';
+                return $this->sendError([$procurementCatCodeExist['id'], $errorMessage], 409);
+            }
         }
 
-        $procurementCatDesExist = TenderProcurementCategory::select('id')
+        $procurementCatDesExist = TenderProcurementCategory::withTrashed()
+            ->select('id', 'deleted_at')
             ->where('description', '=', $input['description'])
             ->where('level', '=', $level)->first();
         if (!empty($procurementCatDesExist)) {
-            return $this->sendError('Procurement category description ' . $input['description'] . ' already exists');
+            if(is_null($procurementCatDesExist['deleted_at'])){
+                return $this->sendError('Procurement category description ' . $input['description'] . ' already exists');
+            } else {
+                $errorMessage = 'Procurement category description '. $input['description'] . ' already exist, Do you really want to restore this?';
+                return $this->sendError([$procurementCatDesExist['id'], $errorMessage], 409);
+            }
         }
 
         $input['created_pc'] = gethostname();
@@ -117,6 +131,10 @@ class TenderProcurementCategoryController extends AppBaseController
 
         $input = $request->all();
 
+        if(isset($input['restore']) && $input['restore'] == true){
+            return $this->restoreDeletedCategory($id);
+        }
+
         if(isset($input['level'])){
             $level = $input['level'];
         }
@@ -141,24 +159,36 @@ class TenderProcurementCategoryController extends AppBaseController
             return $this->sendError($validator->messages(), 422);
         }
 
-        $procurementCodeExist = TenderProcurementCategory::select('id')
+        $procurementCodeExist = TenderProcurementCategory::withTrashed()
+            ->select('id', 'deleted_at')
             ->where('id', '!=', $id)
             ->where('code', '=', $input['code'])
             ->where('level', '=', $level)->first();
         if (!empty($procurementCodeExist)) {
-            return $this->sendError('Procurement code ' . $input['code'] . ' already exists');
+            if(is_null($procurementCodeExist['deleted_at'])) {
+                return $this->sendError('Procurement code ' . $input['code'] . ' already exists');
+            } else {
+                $errorMessage = 'Procurement code '. $input['code'] . '  already exist, Do you really want to restore this?';
+                return $this->sendError([$procurementCodeExist['id'], $errorMessage], 409);
+            }
         }
 
-        $procurementDesExist = TenderProcurementCategory::select('id')
+        $procurementDesExist = TenderProcurementCategory::withTrashed()
+            ->select('id', 'deleted_at')
             ->where('id', '!=', $id)
             ->where('description', '=', $input['description'])
             ->where('level', '=', $level)->first();
         if(!empty($procurementDesExist)) {
-            return $this->sendError('Procurement category description ' . $input['description'] . ' already exists');
+            if(is_null($procurementCodeExist['deleted_at'])) {
+                return $this->sendError('Procurement category description ' . $input['description'] . ' already exists');
+            } else {
+                $errorMessage = 'Procurement category description '. $input['description'] . ' already exist, Do you really want to restore this?';
+                return $this->sendError([$procurementCodeExist['id'], $errorMessage], 409);
+            }
         }
 
-        $input['created_pc'] = gethostname();
-        $input['created_by'] = Helper::getEmployeeID();
+        $input['updated_pc'] = gethostname();
+        $input['updated_by'] = Helper::getEmployeeID();
         $input['parent_id'] = 0;
         $input['level'] = $level;
         $input['parent_id'] = $parent_id;
@@ -181,8 +211,12 @@ class TenderProcurementCategoryController extends AppBaseController
         if (empty($tenderProcurementCategory)) {
             return $this->sendError('Procurement Category not found');
         }
-
-        $tenderProcurementCategory->delete();
+        $input['deleted_by'] = Helper::getEmployeeID();
+        $procurementCategoryDeleted = TenderProcurementCategory::where('id', $id)->update($input);
+        
+        if($procurementCategoryDeleted){
+            $tenderProcurementCategory->delete();
+        }
 
         return $this->sendResponse($id, 'Procurement category deleted successfully');
     }
@@ -224,5 +258,12 @@ class TenderProcurementCategoryController extends AppBaseController
             ->addColumn('Actions', 'Actions', "Actions")
             ->addIndexColumn()
             ->make(true);
+    }
+
+    private function restoreDeletedCategory($id){
+
+        $procurementCategory = TenderProcurementCategory::withTrashed()->find($id)->restore();
+
+        return $this->sendResponse($procurementCategory, 'Procurement Category updated successfully');
     }
 }
